Isofox: record base depth during read assignment and use for junction depth

+++ isofox/src/main/java/com/hartwig/hmftools/isofox/common/ReadRecord.java
@@ -525,8 +525,8 @@ public class ReadRecord
                 continue;
 
             // process this overlap
-            int regionBaseIndex = readStartPos > region.start() ? (int)(readStartPos - region.start()) : 0;
-            int overlap = (int)(min(readEndPos, region.end()) - max(readStartPos, region.start())) + 1;
+            int regionBaseIndex = readStartPos > region.start() ? readStartPos - region.start() : 0;
+            int overlap = min(readEndPos, region.end()) - max(readStartPos, region.start()) + 1;
 
             if(regionBaseIndex + overlap > regionBaseDepth.length)
             {
